Some work on the tests to work for other people.
I paramaterized the user values used in the test so that one can set
values in .env to use different users, projects, and email addresses.

Also moved most of the code in the 2lo test into a test method so
visual studio code's test detection can detect the test, and does not
fail while running the introspection code because of everything
at the file level.

Also added an explicit pytest dependency so uv works for running
pytest in addition to poetry.

This is:
```
  uv run pytest tests/
```

The deletion part of the test still doesn't work for me, but I assume
that I need some differeng test project setup.  Everything else works
as before and now tests can be ran inside of vscode, cursor, etc.